requirements.txt
Pyrogram==2.0.106
tgcrypto
aiosqlite
python-dotenv
requests